codes/requirements.laxed.txt
# Fundamentals
numpy
pyyaml
tb-nightly
future
scp
tqdm
matplotlib
scipy
munch
tqdm
scp
tensorboard
orjson
einops
lambda-networks
mup

#UI
customtkinter
ruamel.yaml
# For image generation stuff
opencv-python
kornia
pytorch_ssim
gsa-pytorch
pytorch_fid

# For audio generation stuff
inflect
librosa
Unidecode
tgt
pyworld
audio2numpy
SoundFile
arabic-reshaper
python-bidi
# For text stuff
transformers==4.29.2
tokenizers
jiwer  # calculating WER
omegaconf

# lucidrains stuff
vector_quantize_pytorch
linear_attention_transformer
rotary-embedding-torch
axial_positional_embedding
g-mlp-pytorch
x-clip
x_transformers==1.0.4

# bitsandbytes
bitsandbytes
lion-pytorch==0.0.7
# triton==2.0.0a2
